backend/pyproject.toml
[tool.poetry]
name = "online-shop-backend"
version = "0.1.0"
description = ""
authors = ["Екатерина Хиталенко"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
